Split out Demo from App
For hot reloading to work, the rendered component must be exported
from an individual module
